fix: prevented duplicate bump item from being created
when adding a new bump, checked if the 'offer product' already exists in another bump with the same 'bump schedule'. if yes, then checked if any of the selected 'categories' or 'products' exists in that 'bump item' and if exist then showed warning and prevented the submission there. Otherwise Proceeded the submission and added the bump item to the bump list.

Resolves #30