fix: when same 'offer product' with different 'offer price' needs to be shown, made sure the best offer among them is shown
when there's multiple bumps for same products with different 'offer price' and they all are eligible to show in the checkout page, made sure the one with be 'best offer' (cheapest price) show's for that offer product.

Before this fix, in the described scenario the bump that was created last was showing in the checkout page for that 'offer product' regardless of their 'offer price', which shouldn't be the case.

For more clarity, take 2 bumps for example, where both of the bump is for a product 'smartphone'. One of the bump is like: 'if you buy a 'laptop' you'll get '30% discount' on the 'smartphone'. And the other one is, if you buy a 'macbook' you'll get '50% discount' on the 'smartphone'.
Now, if a customer adds both 'laptop' & 'macbook' to their cart which one of 'those 2 discounts for the smartphone' should be shown? should he see the '30%' or should he see the '50%' discount? Clearly he should get the 'best' among those 2 which is '50%'. And that's basically what was done in this commit.

Resolves #47