fix(sales pop)!: mobile view not working
Before the fix it was unable to control the view for mobile devices even though the mobile device vew was turned off it was still showing the pop up in the mobile devices. After the fix when the view for mobile device is turned off the pop up will not show in the mobile device except desktop devices.

Resolves: #26