fix(sales-pop): multiple setting not working
Before the fix of the sales pop there are a lot of functionalities that are not woking properly. These settingso fields are popup width, border radius of image, image spacing, link image  to product page and product link in new tab. Which are not working properly. After givng this fixes this issues are being fixed and it is working properly.

Resolves: #15