feat(admin-ui): dynamic pagination and page view data
**Description**
Before the the fix the pagination was only suitable for upto page of 2 but if the page exceeds more than 2 the module content was not viewing properly.

After this feature implementation the page data is dynamically changeable if page exceeds more than 2 the pagination and page data view will work fine.

resolves: #194